Profile Testing

+++ crowdfunding/users/views.py
@@ -1,13 +1,46 @@
 from django.http import Http404
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status, permissions
-from .models import CustomUser
-from .serializers import CustomUserSerializer
-# from .permissions import IsOwnerOrReadOnly
+from rest_framework import status, permissions, generics, viewsets
+from .models import CustomUser, Profile
+from .serializers import CustomUserSerializer, ProfileSerializer
+from .permissions import IsOwnerOrReadOnly, isSuperUser
+from django.shortcuts import get_object_or_404
+from django.core.exceptions import PermissionDenied
 
+# https://stackoverflow.com/questions/55913920/django-rest-framework-displaying-the-users-profile
+
+# class Profile(APIView):
+#     # serializer_class = ProfileSerializer
+
+#     def get(self, request, *args, **kwargs):
+#         user = get_object_or_404(CustomUser, pk=kwargs['user_id'])
+#         print (user)
+#         # print (user.userprofile)
+
+#         profile_serializer = ProfileSerializer(user.userprofile)
+#         return Response(profile_serializer.data)
+    
+    # def get_object(self, pk):
+    #     try:
+    #         return Projects.objects.get(pk=pk)
+    #     except Projects.DoesNotExist:
+    #         raise Http404
+    
+
+    # def get_queryset(self):
+    #     return Profile.objects.filter(user=self.kwargs['user_id'])
+
+    # def get_queryset(self):
+    #     """
+    #     This view should return a list of all the purchases
+    #     for the currently authenticated user.
+    #     """
+    #     user = self.request.user
+    #     return Projects.objects.filter(owner__username=user)
 
 class CustomUserList(APIView):
+    # permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def get(self, request):
         users = CustomUser.objects.all()
@@ -17,12 +50,12 @@ class CustomUserList(APIView):
     def post(self, request):
         serializer = CustomUserSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save(username=request.user)
+            serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors)
 
 class CustomUserDetail(APIView):
-    # permission_classes = [permissions.IsAuthenticatedOrReadOnly]
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly,IsOwnerOrReadOnly, isSuperUser]
 
     def get_object(self, pk):
         try:
@@ -38,7 +71,7 @@ class CustomUserDetail(APIView):
 
     def put(self, request, pk):
         user = self.get_object(pk)
-        # self.check_object_permissions(request, user)
+        self.check_object_permissions(request, user)
         data = request.data
         serializer = CustomUserSerializer(
             instance=user,
@@ -58,10 +91,19 @@ class CustomUserDetail(APIView):
     
     def delete(self, request, pk):
         user = self.get_object(pk)
-        # self.check_object_permissions(request, user)
+        self.check_object_permissions(request, user)
 
         try:
             user.delete()
             return Response(status = status.HTTP_204_NO_CONTENT)
         except CustomUser.DoesNotExist:
             raise Http404
+# raise PermissionDenied()
+
+
+
+# class Profile(viewsets.ViewSet):
+#     def get(self, request, *args, **kwargs):
+#         user = get_object_or_404(CustomUser, pk=kwargs['user_id'])
+#         profile_serializer = ProfileSerializer(user.profile)
+#         return Response(profile_serializer.data)
